+++ src/App.js
@@ -37,6 +37,7 @@ import {
 import { MainThemeProvider } from './context/ThemeContext';
 import { InitialRouteContext } from './context/initialRoute';
 import monitorNetwork from './debugging/network';
+import appEvents from './handlers/appEvents';
 import handleDeeplink from './handlers/deeplinks';
 import { runWalletBackupStatusChecks } from './handlers/walletReadyEvents';
 import RainbowContextWrapper from './helpers/RainbowContext';
@@ -45,9 +46,9 @@ import * as keychain from './model/keychain';
 import { loadAddress } from './model/wallet';
 import { Navigation } from './navigation';
 import RoutesComponent from './navigation/Routes';
+import { explorerInit } from './redux/explorer';
 import { requestsForTopic } from './redux/requests';
 import store from './redux/store';
-import { walletConnectLoadState } from './redux/walletconnect';
 import Routes from '@rainbow-me/routes';
 import logger from 'logger';
 import { Portal } from 'react-native-cool-modals/Portal';
@@ -87,6 +88,7 @@ class App extends Component {
     }
     this.identifyFlow();
     AppState.addEventListener('change', this.handleAppStateChange);
+    appEvents.on('transactionConfirmed', this.handleTransactionConfirmed);
     await this.handleInitializeAnalytics();
     saveFCMToken();
     this.onTokenRefreshListener = registerTokenRefreshListener();
@@ -215,11 +217,6 @@ class App extends Component {
   };
 
   handleAppStateChange = async nextAppState => {
-    // Restore WC connectors when going from BG => FG
-    if (this.state.appState === 'background' && nextAppState === 'active') {
-      store.dispatch(walletConnectLoadState());
-    }
-
     this.setState({ appState: nextAppState });
 
     analytics.track('State change', {
@@ -230,6 +227,14 @@ class App extends Component {
 
   handleNavigatorRef = navigatorRef =>
     Navigation.setTopLevelNavigator(navigatorRef);
+
+  handleTransactionConfirmed = () => {
+    logger.log('Reloading all data from zerion in 10!');
+    setTimeout(() => {
+      logger.log('Reloading all data from zerion NOW!');
+      store.dispatch(explorerInit());
+    }, 10000);
+  };
 
   render = () => (
     <MainThemeProvider>