+++ src/handlers/appEvents.js
@@ -1,0 +1,5 @@
+import { EventEmitter } from 'events';
+
+const appEvents = new EventEmitter();
+
+export default appEvents;