+++ src/hooks/index.js
@@ -101,5 +101,5 @@ export { default as useForceUpdate } from './useForceUpdate';
 export { default as useOnAvatarPress } from './useOnAvatarPress';
 export { default as usePoolDetails } from './usePoolDetails';
 export { default as useAdditionalAssetData } from './useAdditionalAssetData';
-export { useTotalFeeEarnedPerAsset } from './useUsersPositions';
+export { default as useTotalFeeEarnedPerAsset } from './useTotalFeeEarnedPerAsset';
 export { default as useImportingWallet } from './useImportingWallet';