+++ src/hooks/useGas.js
@@ -49,19 +49,20 @@ export default function useGas() {
   );
 
   const updateDefaultGasLimit = useCallback(
-    defaultGasLimit => dispatch(gasUpdateDefaultGasLimit(defaultGasLimit)),
+    (network, defaultGasLimit) =>
+      dispatch(gasUpdateDefaultGasLimit(network, defaultGasLimit)),
     [dispatch]
   );
 
   const updateGasPriceOption = useCallback(
-    (option, network = currentNetwork) =>
-      dispatch(gasUpdateGasPriceOption(option, network)),
+    (option, network = currentNetwork, assetsOverride = null) =>
+      dispatch(gasUpdateGasPriceOption(option, network, assetsOverride)),
     [currentNetwork, dispatch]
   );
 
   const updateTxFee = useCallback(
     (newGasLimit, overrideGasOption, network = currentNetwork) => {
-      dispatch(gasUpdateTxFee(newGasLimit, overrideGasOption, network));
+      dispatch(gasUpdateTxFee(network, newGasLimit, overrideGasOption));
     },
     [currentNetwork, dispatch]
   );