+++ src/model/migrations.js
@@ -27,10 +27,11 @@ import {
   getUserLists,
   saveUserLists,
 } from '@rainbow-me/handlers/localstorage/userLists';
+import { resolveNameOrAddress } from '@rainbow-me/handlers/web3';
 import { returnStringFirstEmoji } from '@rainbow-me/helpers/emojiHandler';
 import { updateWebDataEnabled } from '@rainbow-me/redux/showcaseTokens';
 import { DefaultTokenLists } from '@rainbow-me/references';
-import { defaultProfileUtils } from '@rainbow-me/utils';
+import { profileUtils } from '@rainbow-me/utils';
 import logger from 'logger';
 
 export default async function runMigrations() {
@@ -311,16 +312,14 @@ export default async function runMigrations() {
           return {
             ...account,
             ...(!accountEmoji && {
-              label: `${defaultProfileUtils.addressHashedEmoji(
-                account.address
-              )} ${account.label}`,
+              label: `${profileUtils.addressHashedEmoji(account.address)} ${
+                account.label
+              }`,
             }),
             color:
               (accountEmoji
                 ? newColorIndexes[account.color]
-                : defaultProfileUtils.addressHashedColorIndex(
-                    account.address
-                  )) || 0,
+                : profileUtils.addressHashedColorIndex(account.address)) || 0,
           };
         });
         const newWallet = { ...wallet, addresses: newAddresses };
@@ -364,6 +363,53 @@ export default async function runMigrations() {
   };
 
   migrations.push(v9);
+
+  /*
+   *************** Migration v10 ******************
+   * This step makes sure all contacts have an emoji set based on the address
+   */
+  const v10 = async () => {
+    logger.log('Start migration v10');
+    try {
+      // migrate contacts to corresponding emoji
+      const contacts = await getContacts();
+      let updatedContacts = { ...contacts };
+      if (!contacts) return;
+      const contactKeys = Object.keys(contacts);
+      for (let j = 0; j < contactKeys.length; j++) {
+        const contact = contacts[contactKeys[j]];
+        let nickname = contact.nickname;
+        if (!returnStringFirstEmoji(nickname)) {
+          let address = null;
+          try {
+            address = await resolveNameOrAddress(contact.address);
+          } catch (error) {
+            const migrationError = new Error(
+              `Error during v10 migration contact address resolution for ${contact.address}`
+            );
+            captureException(migrationError);
+            continue;
+          }
+          const emoji = profileUtils.addressHashedEmoji(address);
+          const color = profileUtils.addressHashedColorIndex(address);
+          nickname = `${emoji} ${nickname}`;
+          updatedContacts[contactKeys[j]] = {
+            ...contact,
+            color,
+            nickname,
+          };
+        }
+      }
+      logger.log('update contacts to add emojis / colors');
+      await saveContacts(updatedContacts);
+    } catch (error) {
+      logger.sentry('Migration v10 failed: ', error);
+      const migrationError = new Error('Migration 10 failed');
+      captureException(migrationError);
+    }
+  };
+
+  migrations.push(v10);
 
   logger.sentry(
     `Migrations: ready to run migrations starting on number ${currentVersion}`