+++ src/utils/ethereumUtils.js
@@ -20,7 +20,8 @@ import { Linking, NativeModules } from 'react-native';
 import { ETHERSCAN_API_KEY } from 'react-native-dotenv';
 import { useSelector } from 'react-redux';
 import URL from 'url-parse';
-import { isTestnet } from '@rainbow-me/handlers/web3';
+import { getOnchainAssetBalance } from '@rainbow-me/handlers/assets';
+import { getProviderForNetwork, isTestnet } from '@rainbow-me/handlers/web3';
 import networkTypes from '@rainbow-me/helpers/networkTypes';
 import {
   convertAmountAndPriceToNativeDisplay,
@@ -39,15 +40,69 @@ import {
 import store from '@rainbow-me/redux/store';
 import {
   ARBITRUM_BLOCK_EXPLORER_URL,
+  ARBITRUM_ETH_ADDRESS,
   chains,
   ETH_ADDRESS,
   MATIC_MAINNET_ADDRESS,
+  MATIC_POLYGON_ADDRESS,
   OPTIMISM_BLOCK_EXPLORER_URL,
+  OPTIMISM_ETH_ADDRESS,
   POLYGON_BLOCK_EXPLORER_URL,
 } from '@rainbow-me/references';
 import logger from 'logger';
 
 const { RNBip39 } = NativeModules;
+
+const getNativeAssetForNetwork = async (network, address) => {
+  let nativeAssetAddress;
+  switch (network) {
+    case networkTypes.arbitrum:
+      nativeAssetAddress = ARBITRUM_ETH_ADDRESS;
+      break;
+    case networkTypes.optimism:
+      nativeAssetAddress = OPTIMISM_ETH_ADDRESS;
+      break;
+    case networkTypes.polygon:
+      nativeAssetAddress = MATIC_POLYGON_ADDRESS;
+      break;
+    default:
+      nativeAssetAddress = ETH_ADDRESS;
+  }
+  const { accountAddress } = store.getState().settings;
+  let differentWallet = toLower(address) !== toLower(accountAddress);
+  const { assets } = store.getState().data;
+  let nativeAsset =
+    !differentWallet && getAsset(assets, toLower(nativeAssetAddress));
+
+  // If the asset is on a different wallet, or not available in this wallet
+  if (differentWallet || !nativeAsset) {
+    const { genericAssets } = store.getState().data;
+    const mainnetAddress =
+      network === networkTypes.polygon ? MATIC_MAINNET_ADDRESS : ETH_ADDRESS;
+    nativeAsset = genericAssets[mainnetAddress];
+
+    if (network === networkTypes.polygon) {
+      nativeAsset.mainnet_address = mainnetAddress;
+      nativeAsset.address = MATIC_POLYGON_ADDRESS;
+    }
+
+    const provider = await getProviderForNetwork(network);
+    const balance = await getOnchainAssetBalance(
+      nativeAsset,
+      address,
+      network,
+      provider
+    );
+
+    const assetWithBalance = {
+      ...nativeAsset,
+      balance,
+    };
+
+    return assetWithBalance;
+  }
+  return nativeAsset;
+};
 
 const getAsset = (assets, address = 'eth') =>
   find(assets, matchesProperty('address', toLower(address)));
@@ -74,6 +129,12 @@ export const useEthUSDPrice = () => {
 
 export const useEthUSDMonthChart = () => {
   return useSelector(({ charts: { chartsEthUSDMonth } }) => chartsEthUSDMonth);
+};
+
+const getPriceOfNativeAssetForNetwork = network => {
+  return network === networkTypes.polygon
+    ? getMaticPriceUnit()
+    : getEthPriceUnit();
 };
 
 const getEthPriceUnit = () => getAssetPrice();
@@ -381,8 +442,10 @@ export default {
   getEthPriceUnit,
   getHash,
   getMaticPriceUnit,
+  getNativeAssetForNetwork,
   getNetworkFromChainId,
   getNetworkNameFromChainId,
+  getPriceOfNativeAssetForNetwork,
   hasPreviousTransactions,
   isEthAddress,
   openAddressInBlockExplorer,